Fixing bracket/parenthesis not found bug
Failure-inducing inputs: Test files 2, 3, 4, 7, and 8 were all fixed by this bug fix

Example symptom:
5) testFile7(MarkdownParseTest)
java.lang.OutOfMemoryError: Java heap space
        at java.base/java.util.Arrays.copyOf(Arrays.java:3512)
        at java.base/java.util.Arrays.copyOf(Arrays.java:3481)
        at java.base/java.util.ArrayList.grow(ArrayList.java:237)
        at java.base/java.util.ArrayList.grow(ArrayList.java:244)
        at java.base/java.util.ArrayList.add(ArrayList.java:454)
        at java.base/java.util.ArrayList.add(ArrayList.java:467)
        at MarkdownParse.getLinks(MarkdownParse.java:18)
        at MarkdownParseTest.testFile7(MarkdownParseTest.java:62)

Bug that fixed symptom:
When parentheses or brackets weren't found, a value of -1 was returned and fed into the substring line, which caused infinite loops and invalid input errors (depending on which was missing). I fixed this by adding an if statement to break if any of the parentheses or brackets could not be found (in other words, the set was incomplete and no more valid links existed).