Fixed bug with characters in-between ']' and '('
Failure-inducing input fixed: test file 5

Symptom:
1) testFile5(MarkdownParseTest)
java.lang.AssertionError: expected:<[page.com]> but was:<[]>
        at org.junit.Assert.fail(Assert.java:89)
        at org.junit.Assert.failNotEquals(Assert.java:835)
        at org.junit.Assert.assertEquals(Assert.java:120)
        at org.junit.Assert.assertEquals(Assert.java:146)
        at MarkdownParseTest.testFile5(MarkdownParseTest.java:48)

Bug that fixed symptom: The bug in the code was that the code could not identify when there were characters in-between the closed bracket and the open parentheses, which caused false positives (since that should invalidate a markdown link). To solve this issue, I added a clause to the if-statement that continues past an invalid set of []() that checked if the character before '(' was ']' and continued past if it wasn't.